rename deprecated module
