+++ src/screens/RoomTextbookPage.tsx
@@ -9,7 +9,7 @@ import DefaultLayout from '../complexes/DefaultLayout';
 import Container from '../independents/Container';
 import LoadingView from '../independents/LoadingView';
 import { store } from '../misc';
-import { connectActiveRooms } from '../models/rooms';
+import { connectActiveRooms } from '../models/rooms0';
 import { IState } from '../reducers';
 import { IUserProfile } from '../reducers/currentUser';
 import { IRoom } from '../reducers/rooms';
